+++ src/test/java/roomescape/util/PaymentClientTestConfiguration.java
@@ -1,4 +1,4 @@
-package roomescape.helper;
+package roomescape.util;
 
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
